remove test state and restructure functions
